api/requirements.txt
certifi==2020.6.20
chardet==3.0.4
Django==2.2.12
django-cors-headers==3.2.1
djangorestframework==3.11.0
idna==2.10
Markdown==3.2.1
psycopg2==2.8.5
python-decouple==3.3
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10